[HCW] Rename ConfiguratorDlg2 -> ConfigWizard.
